split workflows into files
